RAG/requirements.txt
langchain==0.3.3
langchain_openai==0.2.2
langchain_core==0.3.10
python-dotenv==1.0.1
streamlit==1.39.0
langchain_community==0.3.2
langserve==0.3.0
fastapi==0.115.0
uvicorn==0.31.1
sse_starlette==2.1.3
sentence_transformers==3.1.1
chromadb==0.5.12
numpy
pandas
pysqlite3-binary